consolidate component exports and update imports
