+++ src/app/_utils/index.tsx
@@ -49,5 +49,5 @@ export const logoLineGen = (appTheme: string, size: string) => {
 };
 
 export const pagePadding = () => {
-  return `px-8 pt-10 pb-4`;
+  return `px-8 pt-8 pb-4`;
 };
