+++ src/app/components/index.tsx
@@ -1,0 +1,29 @@
+import CategoryList from "./categoryList";
+import CategoryTile from "./categoryTile";
+import Header from "./header";
+import HeaderMenu from "./headerMenu";
+import ListHeader from "./listHeader";
+import Logo from "./logo";
+import ProductDisplay from "./productDisplay";
+import ProductList from "./productList";
+import ProductTile from "./productTile";
+import SearchBar from "./searchBar";
+import Sidebar from "./sideBar";
+import SidebarCap from "./sidebarCap";
+import ToggleSwitch from "./toggleSwitch";
+
+export {
+  CategoryList,
+  CategoryTile,
+  Header,
+  HeaderMenu,
+  ListHeader,
+  Logo,
+  ProductDisplay,
+  ProductList,
+  ProductTile,
+  SearchBar,
+  Sidebar,
+  SidebarCap,
+  ToggleSwitch,
+};
