+++ src/app/components/listHeader.tsx
@@ -4,7 +4,7 @@ type ListHeaderProps = {
   title: string;
 };
 
-export function ListHeader(props: ListHeaderProps) {
+export default function ListHeader(props: ListHeaderProps) {
   const { title } = props;
   return (
     <div className={`${lobsterFont.className} mb-6 text-5xl`}>{title}</div>
