+++ src/app/components/sidebarCap.tsx
@@ -5,9 +5,9 @@ import { useContext } from "react";
 import { themeGen } from "../_utils";
 import Link from "next/link";
 import { lobsterFont } from "../fonts";
-import { Logo } from "./logo";
+import { Logo } from "./index";
 
-export function SidebarCap() {
+export default function SidebarCap() {
   const { appTheme } = useContext(ThemeContext);
 
   return (
